+++ lib/hubspot/config.rb
@@ -31,6 +31,10 @@ module Hubspot
           raise Hubspot::ConfigurationError.new("You must provide either an access_token or an hapikey")
         end
 
+        if hapikey.present?
+          Hubspot::Deprecator.build.deprecation_warning("hapikey", "please use access_token instead. See https://developers.hubspot.com/docs/api/migrate-an-api-key-integration-to-a-private-app")
+        end
+
         if access_token.present?
           Hubspot::Connection.headers("Authorization" => "Bearer #{access_token}")
         end