+++ spec/lib/hubspot/custom_event_spec.rb
@@ -1,7 +1,8 @@
 RSpec.describe Hubspot::CustomEvent do
-  let(:portal_id) { '62515' }
-  let(:sent_portal_id) { portal_id }
-  before { Hubspot.configure(hapikey: 'demo', custom_event_prefix: 'foobar') }
+  before do
+    Hubspot.configure(access_token: ENV.fetch("HUBSPOT_ACCESS_TOKEN"), portal_id: ENV.fetch("HUBSPOT_PORTAL_ID"),
+                      custom_event_prefix: 'foobar')
+  end
 
   describe '.trigger' do
     let(:event_name) { 'my_awesome_event' }
@@ -9,7 +10,7 @@ RSpec.describe Hubspot::CustomEvent do
     let(:properties) { { prop_foo: 'bar' } }
     let(:options) { {} }
     let(:base_url) { 'https://api.hubapi.com' }
-    let(:url) { "#{base_url}/events/v3/send?hapikey=demo" }
+    let(:url) { "#{base_url}/events/v3/send" }
 
     subject { described_class.trigger(event_name, email, properties, options) }