+++ spec/lib/hubspot/event_spec.rb
@@ -1,7 +1,8 @@
 describe Hubspot::Event do
-  let(:portal_id) { '62515' }
+  let(:portal_id) { ENV.fetch("HUBSPOT_PORTAL_ID") }
   let(:sent_portal_id) { portal_id }
-  before { Hubspot.configure(hapikey: 'demo', portal_id: portal_id) }
+
+  before { Hubspot.configure(access_token: ENV.fetch("HUBSPOT_ACCESS_TOKEN"), portal_id: portal_id) }
 
   describe '.trigger' do
     let(:event_id) { '000000001625' }