+++ spec/lib/hubspot/file_spec.rb
@@ -1,12 +1,23 @@
-describe Hubspot do
-  describe Hubspot::File do
-    before { Hubspot.configure(hapikey: "demo") }
-    describe ".find_by_id" do
-      it "should fetch specific file" do
-        VCR.use_cassette("file_find") do
-          file = Hubspot::File.find_by_id(5323360053)
-          file.id.should eq(5323360053)
-        end
+describe Hubspot::File do
+  let(:example_file_hash) do
+    VCR.use_cassette('file_example') do
+      headers = { Authorization: "Bearer #{ENV.fetch('HUBSPOT_ACCESS_TOKEN')}" }
+      body = {
+        file: File.open(File::NULL, 'r'),
+        options: { access: 'PRIVATE', ttl: 'P1M', overwrite: false }.to_json,
+        folderPath: '/'
+      }
+      url = 'https://api.hubapi.com/filemanager/api/v3/files/upload'
+      HTTParty.post(url, body: body, headers: headers).parsed_response
+    end
+  end
+
+  describe ".find_by_id" do
+    it "should fetch specific file" do
+      VCR.use_cassette("file_find") do
+        id = example_file_hash['objects'].first['id']
+        file = Hubspot::File.find_by_id(id)
+        file.id.should eq(id)
       end
     end
   end