+++ spec/support/vcr.rb
@@ -10,4 +10,5 @@ VCR.configure do |c|
   c.default_cassette_options = { record: vcr_record_mode }
   c.filter_sensitive_data("<HAPI_KEY>") { ENV.fetch("HUBSPOT_HAPI_KEY") }
   c.filter_sensitive_data("<PORTAL_ID>") { ENV.fetch("HUBSPOT_PORTAL_ID") }
+  c.filter_sensitive_data("<ACCESS_TOKEN>") { ENV.fetch("HUBSPOT_ACCESS_TOKEN") }
 end